# src/main/resources/openapi/gymflex_api_v1.yaml
openapi: 3.0.1
info:
  description: "This project is the backend of the GymFlex application. It is a RESTful API that provides endpoints for the GymFlex frontend to interact with the database."
  title: GymFlexApi
  version: 1.0.0
servers:
  - url: http://localhost:8081/api
    description: Local server
security:
  - bearerAuth: [ ]
...
paths:
...
  /online-users:
    post:
      tags:
        - User
      summary: "Create a new online user"
      operationId: "createOnlineUser"
      requestBody:
        description: "The online user creation request body"
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OnlineUserDTO'
      responses:
        '201':
          description: "Online user created successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnlineUserDTO'
        '400':
          description: "Bad request"
        '500':
          description: "Server error"
    get:
      tags:
        - User
      summary: "Get all online users"
      operationId: "getAllOnlineUsers"
      responses:
        '200':
          description: "List of all online users"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OnlineUserDTO'
        '500':
          description: "Server error"
  /online-users/{id}:
    get:
      tags:
        - User
      summary: "Get online user by ID"
      operationId: "getOnlineUser"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: "Online user retrieved successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnlineUserDTO'
        '404':
          description: "Online user not found"
        '500':
          description: "Server error"
    put:
      tags:
        - User
      summary: "Update an existing online user"
      operationId: "updateOnlineUser"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        description: "The online user update request body"
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OnlineUserDTO'
      responses:
        '200':
          description: "Online user updated successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnlineUserDTO'
        '400':
          description: "Bad request"
        '404':
          description: "Online user not found"
        '500':
          description: "Server error"
    patch:
      tags:
        - User
      summary: "Partial update of an online user"
      operationId: "partialUpdateOnlineUser"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        description: "The online user partial update request body"
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OnlineUserDTO'
      responses:
        '200':
          description: "Online user updated successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnlineUserDTO'
        '400':
          description: "Bad request"
        '404':
          description: "Online user not found"
        '500':
          description: "Server error"
    delete:
      tags:
        - User
      summary: "Delete an online user"
      operationId: "deleteOnlineUser"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: "Online user deleted successfully"
        '404':
          description: "Online user not found"
        '500':
          description: "Server error"
  /online-users/upload/{name}:
    post:
      tags:
        - User
      summary: "Upload member profile picture"
      operationId: "handleFileUpload"
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: "The file upload request body"
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: "File uploaded successfully"
        '500':
          description: "Server error"
  /online-users/upload/{id}:
    put:
      tags:
        - User
      summary: "Update member profile picture"
      operationId: "updateFileUpload"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        description: "The file upload request body"
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: "File updated successfully"
        '500':
          description: "Server error"
  /web/contact:
    post:
      summary: "Handle contact requests"
      operationId: "contact"
      requestBody:
        description: "The contact email details"
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactMailDTO'
      responses:
        '200':
          description: "Email sent successfully"
        '500':
          description: "Server error"
  /web/verify/{validationkey}:
    get:
      summary: "Verify email address"
      operationId: "verifyEmail"
      parameters:
        - name: validationkey
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: "Email verified successfully"
        '500':
          description: "Server error"
...
  /equipment/upload/{name}:
    post:
      tags:
        - Equipment
      summary: Handle file upload for equipment
      operationId: handleFileUpload
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: The file to upload
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: File uploaded successfully
        '500':
          description: Server error
  /equipment/upload/{id}:
    put:
      tags:
        - Equipment
      summary: Update file upload for equipment
      operationId: updateFileUpload
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        description: The file to upload
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: File updated successfully
        '500':
          description: Server error
components:
  schemas:
...
    RoleDTO:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        analytics:
          type: boolean
        membership:
          type: boolean
        payments:
          type: boolean
        inventory:
          type: boolean
        training:
          type: boolean
        settings:
          type: boolean
        preferences:
          type: boolean
        manageWebSite:
          type: boolean
        coach:
          type: boolean
    OnlineUserDTO:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        password:
          type: string
        login:
          type: string
        profilePicture:
          type: string
        roles:
          $ref: '#/components/schemas/RoleDTO'
      required:
        - firstName
        - lastName
        - email
        - password
        - login
    ContactMailDTO:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        subject:
          type: string
        message:
          type: string
      required:
        - name
        - email
        - subject
        - message
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT